requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
tensorflow-cpu